accountBalance.js
Program calculates the total amount of the deposits in a bank account. It reads the amount from a source. When it meets "NoMoreMoney" the program stops and writes the total amount of the money in the bank account. When it receives a negative number from the source it writes "Invalid operation!".