graduation.js
Program calculates the average rating of a student through his curriculum. It receives the name of the student on the first place from the source. On the other next lines it receives the annual ratings of the student. If his annual rate is lower than 4.00 he is excluded. If he is excluded more than twice he will be excluded from the school and the program is over. In this case the program writes the name and the grade of the student. If the student reach the final 12 grade he is a graduate and the program is over. In this case the program writes the name of the student and his average rating for the hole curriculum.