oldBooks.js
Program receives a books from the list and check it with a choose book title. If one of the books from the source match the chosen one. Program writes the amount of the checked books. If there are a string "No More Books" in the list, the program writes "The book you search is not here!" and also "You checked {amount of the checked books} books."