examPreparation.js
Program receive information about student tasks and the rates per tasks on the new line. The first number is the amount of bad rates which he could afford before the program breaks. Bad rates are every single rate lower than or equal of 4. If the program receive a command "Enough", it breaks.