equalArrays.js
Write a program, which receives two string arrays containing number representations, and prints on the console whether they are identical.
Arrays are identical if their elements at same indexes are equal. If they are identical, find the sum of the first array and print the following message: 
`Arrays are identical. Sum: {sum}`
 If the arrays are NOT identical, find the first index where the arrays differ and print the following message:
 `Arrays are not identical. Found difference at {index} index`
Examples
Input	Output
['10','20','30'], ['10','20','30']	Arrays are identical. Sum: 60
['1','2','3','4','5'], ['1','2','4','4','5']	Arrays are not identical. Found difference at 2 index
['1'], ['10']	Arrays are not identical. Found difference at 0 index