equalSums2.js
Write a function that determines if there exists an element in the array of numbers such that the sum of the elements on its left is equal to the sum of the elements on its right. 
If there are NO elements to the left/right, their sum is 0. 
Print the index that satisfies the required condition or "no" if there is no such index.


Input: [1, 2, 3, 3]   Output: 2   
Input: [1, 2]  Output: no 
Input: [1]  Output: 0 
Input: [1, 2, 3]  Output: no
Input: [10, 5, 5, 99, 3, 4, 2, 5, 1, 1, 4]  Output: 3