perfectNumber.js
Write a function that receives a number and checks if that number is perfect or NOT.
A perfect number is a positive integer that is equal to the sum of its proper positive divisors. That is the sum of its positive divisors excluding the number itself (also known as its aliquot sum).
Output
•	If the number is perfect, print: "We have a perfect number!"
•	Otherwise, print: "It's not so perfect."



Input: 6
Output : We have a perfect number!
Comments : 1 + 2 + 3



Input: 28
Output : We have a perfect number!
Comments : 1 + 2 + 4 + 7 + 14



Input: 1236498
Output : It's not so perfect.