Create lastNumbersSequence3.js
You are given two integers n and k. Write a function that generates and prints the following sequence:
•	The first element is 1.
•	Every following element equals the sum of the previous k elements.
•	The length of the sequence is n elements.
	
The input comes as two number arguments. The first element represents the number n, and the second – the number k.
The output is printed on the console on a single line, separated by space.

Input: 6, 3 
Output: 1 1 2 4 7 13

Input: 8, 2 
Output: 1 1 2 3 5 8 13 21