Create train3.js
You will be given an array of strings.
The first element will be a string containing wagons (numbers). Each number inside the string represents the number of passengers that are currently in a wagon. 
The second element in the array will be the max capacity of each wagon (single number).
The rest of the elements will be commands in the following format:
•	Add {passengers} – add a wagon to the end with the given number of passengers.
•	{passengers} - find an existing wagon to fit all the passengers (starting from the first wagon)
At the end, print the final state of the train (all the wagons separated by a space).

Input: ['32 54 21 12 4 0 23',
'75',
'Add 10',
'Add 0',
'30',
'10',
'75']
Output: 72 54 21 12 4 75 23 10 0

Input: ['0 0 0 10 2 4',
'10',
'Add 10',
'10',
'10',
'10',
'8',
'6']
Output: 10 10 10 10 10 10 10